+++ packages/account-v2/src/types.ts
@@ -1,6 +1,11 @@
-import { useKycAuthStatus, useSettings } from '@deriv/api-v2';
-import { CurrencyConstants } from '@deriv-com/utils';
-import { IDV_ERROR_CODES, POI_SERVICE } from './constants';
+import { useGetAccountStatus, useKycAuthStatus, useSettings } from '@deriv/api-v2';
+import {
+    AUTH_STATUS_CODES,
+    getPaymentMethodsConfig,
+    IDV_ERROR_CODES,
+    PAYMENT_METHOD_IDENTIFIER,
+    POI_SERVICE,
+} from './constants';
 
 export type TSupportedDocuments = Exclude<
     Exclude<ReturnType<typeof useKycAuthStatus>['kyc_auth_status'], undefined>['identity']['supported_documents'],
@@ -19,7 +24,6 @@ export type TPOIStatus = Exclude<
 
 export type TGetSettingsResponse = ReturnType<typeof useSettings>['data'];
 
-export type TCurrency = CurrencyConstants.Currency;
 export type TAuthStatusCodes = typeof AUTH_STATUS_CODES[keyof typeof AUTH_STATUS_CODES];
 
 export type TPaymentMethod = keyof ReturnType<typeof getPaymentMethodsConfig>;
@@ -58,3 +62,13 @@ export type TProofOfOwnershipFormValue = Record<TPaymentMethod, Record<number | 
 export type TPOIService = typeof POI_SERVICE[keyof typeof POI_SERVICE];
 
 export type TIDVErrorStatusCode = typeof IDV_ERROR_CODES[keyof typeof IDV_ERROR_CODES]['code'];
+
+export type TAccountLimitValues = {
+    category?: string;
+    hintInfo?: string;
+    isLessProminent?: boolean;
+    title?: string;
+    value?: number | string;
+};
+
+export type TCurrency = CurrencyConstants.Currency;
