+++ packages/account-v2/src/utils/accountLimitsUtils.ts
@@ -1,20 +1,12 @@
 import { GetLimits } from '@deriv/api-types';
 import { FormatUtils } from '@deriv-com/utils';
-import { TCurrency } from '../types';
+import { TAccountLimitValues, TCurrency } from '../types';
 
 export const CATEGORY = {
     footer: 'footer',
     header: 'header',
     submarket: 'submarket',
 } as const;
-
-export type TAccountLimitValues = {
-    category?: string;
-    hintInfo?: string;
-    isLessProminent?: boolean;
-    title?: string;
-    value?: number | string;
-};
 
 type TMarketSpecific = GetLimits['market_specific'];
 type TMarketSpecificData = Exclude<TMarketSpecific, undefined>[string];
